Added SymfonyDriver#getResponseHeaders
It happens that when in a symfony2 context, the client used is
FrameworkBundle\Client, which differs from the Goutte\Client by the Response
object it uses (HttpFoundation\Response instead of BrowserKit\Response).
Unfortunately, those 2 response objects are not compatible.

This implementation might look a bit messy, but sf components' HeaderBag does
not provide a clean way to retrieve headers in the format we want.